Added TypeScript to the JS project.
This may be the wrong direction, however LangChainJS seems to use TypeScript as do the examples. Therefore, that is my plan as well. So, this PR is focused on getting TypeScript setup.

For this PR, I have included a basic retriever for both TypeScript and standard JavaScript. Both of them basic retrievers that return an array of static documents. Those documents are not yet used in the prompt to the LLM.

We'll need to decide whether to make the example TypeScript or JavaScript.